Move log_reader implementation to header and make everything inline
