+++ sa/CMakeLists.txt
@@ -374,7 +374,6 @@ set(TRANSACTION_SOURCES
   ${TRANSACTION_DIR}/log_page_buffer.c
   ${TRANSACTION_DIR}/log_postpone_cache.cpp
   ${TRANSACTION_DIR}/log_prior_send.cpp
-  ${TRANSACTION_DIR}/log_reader.cpp
   ${TRANSACTION_DIR}/log_recovery.c
   ${TRANSACTION_DIR}/log_recovery_analysis.cpp
   ${TRANSACTION_DIR}/log_recovery_context.cpp
